Refactor stdio functions to use size_t for buffer sizes and update kernel to utilize stdio functions for output

+++ include/common/stdio.h
@@ -2,9 +2,10 @@
 #define COMMON_STDIO_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 int32_t getc(void);
-char *gets(char *buf, int32_t size);
+char *gets(char *buf, size_t size);
 void putc(char c);
 void puts(const char *str);
